# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tire Management System API",
        "description": "API documentation for the Tire Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/operations/mount": {
            "post": {
                "tags": [
                    "Operations"
                ],
                "summary": "Mounts a tire onto a vehicle at a specific position Mounts Validates that the tire exists, is not retired or defective, is compatible with the vehicle/axle, and that the vehicle exists. Captures vehicle mileage, technician (logged-inuser), timestamp, optional photos and notes. Updates tire status to mounted and records lifecycle history",
                "operationId": "55c467a4e47c591cfcf8cdff069b5d6f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tire_id",
                                    "vehicle_id",
                                    "position",
                                    "odometer"
                                ],
                                "properties": {
                                    "tire_id": {
                                        "type": "integer"
                                    },
                                    "vehicle_id": {
                                        "type": "integer"
                                    },
                                    "position": {
                                        "type": "string"
                                    },
                                    "odometer": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tire mounted"
                    }
                }
            }
        },
        "/api/operations/dismount": {
            "post": {
                "tags": [
                    "Operations"
                ],
                "summary": "Dismount a tire",
                "operationId": "d36fdc9e60c095f6c49a29a073124a51",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tire_id",
                                    "to_warehouse_id",
                                    "odometer"
                                ],
                                "properties": {
                                    "tire_id": {
                                        "type": "integer"
                                    },
                                    "to_warehouse_id": {
                                        "type": "integer"
                                    },
                                    "odometer": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tire dismounted"
                    }
                }
            }
        },
        "/api/operations/rotate": {
            "post": {
                "tags": [
                    "Operations"
                ],
                "summary": "Swap tire positions",
                "operationId": "faa34a465d0fc445805d5d90243b6c11",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicle_id",
                                    "rotations",
                                    "odometer"
                                ],
                                "properties": {
                                    "vehicle_id": {
                                        "type": "integer"
                                    },
                                    "rotations": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "tire_id": {
                                                    "type": "integer"
                                                },
                                                "new_position": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "odometer": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rotation complete"
                    }
                }
            }
        },
...
        "/api/operations/upload-photo": {
            "post": {
                "tags": [
                    "Operations"
                ],
                "summary": "Upload photo for a tire operation",
                "operationId": "23f37ee59b8cafbf727068690ff35ede",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "operation_id",
                                    "photo"
                                ],
                                "properties": {
                                    "operation_id": {
                                        "type": "integer"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Photo uploaded"
                    }
                }
            }
        },
...
        "/api/stock/tires/{id}": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Get single tire details",
                "operationId": "c0e1a7088bd8a6af5ec1cf6497120f44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tire details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Stock"
                ],
                "summary": "Update tire details",
                "operationId": "7166808fe98e0f32ce28be653258be4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "brand": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "size": {
                                        "type": "string"
                                    },
                                    "cost": {
                                        "type": "number"
                                    },
                                    "vendor": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tire updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Stock"
                ],
                "summary": "Delete a tire",
                "operationId": "2e23389788fb4519a76027eede6c46f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tire deleted"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}